Minor adjustments
`--disable-shared` ensures that you don't fall back to the default of `--enable-shared` for `configure`. Maybe you didn't run into that issue because you ran `cabal configure --disable-executable-dynamic`, and cabal picked up the configure state from there.

The `-pthread` should not be necessary, ghc should add this on it's own when asking for `-threaded` and the system requires libpthread.  You'd end up with `-lpthread` twice in the linker arguments.